requirements.txt
naoqi
python-dotenv